shared/interworks_snowpark/requirements.txt
cryptography
snowflake-snowpark-python[pandas]
snowflake-ingest
streamlit
